refactor(StreetMeta): reorganize files
