+++ assets/scripts/streets/StreetMeta/StreetMeta.scss
@@ -1,5 +1,5 @@
 @use "sass:color";
-@import "../../styles/variables";
+@import "../../../styles/variables";
 
 .street-metadata {
   display: flex;
