+++ assets/scripts/streets/StreetMeta/StreetMeta.test.js
@@ -1,12 +1,8 @@
 /* eslint-env jest */
 import React from 'react'
 import { waitFor } from '@testing-library/react'
-import StreetMeta from '../StreetMeta'
 import { render } from '../../../../test/helpers/render'
-
-jest.mock('../../app/load_resources', () => {})
-jest.mock('../../app/initialization', () => {})
-jest.mock('../../preinit/app_settings', () => {})
+import StreetMeta from './StreetMeta'
 
 describe('StreetMeta', () => {
   it('renders without crashing', async () => {
