+++ assets/scripts/streets/StreetMeta/StreetMetaAnalytics.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { FormattedMessage, useIntl } from 'react-intl'
 import { IoPodiumOutline } from 'react-icons/io5'
-import { useSelector, useDispatch } from '../store/hooks'
-import Tooltip from '../ui/Tooltip'
-import { getStreetCapacity } from '../segments/capacity'
-import { formatNumber } from '../util/number_format'
-import { showDialog } from '../store/slices/dialogs'
+import { useSelector, useDispatch } from '../../store/hooks'
+import { showDialog } from '../../store/slices/dialogs'
+import Tooltip from '../../ui/Tooltip'
+import { getStreetCapacity } from '../../segments/capacity'
+import { formatNumber } from '../../util/number_format'
 
 function StreetMetaAnalytics (): React.ReactElement | null {
   const street = useSelector((state) => state.street)
