+++ assets/scripts/streets/StreetMeta/StreetMetaWidthContainer.test.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import { render } from '../../../../test/helpers/render'
-import StreetMetaWidthContainer from '../StreetMetaWidthContainer'
+import StreetMetaWidthContainer from './StreetMetaWidthContainer'
 
 describe('StreetMetaWidthContainer', () => {
   it('renders', () => {
