+++ assets/scripts/streets/StreetMeta/StreetMetaWidthLabel.scss
@@ -1,5 +1,5 @@
 @use "sass:color";
-@import "../../styles/variables";
+@import "../../../styles/variables";
 
 .street-width {
   display: inline-flex;
