+++ assets/scripts/streets/StreetMeta/StreetMetaWidthLabel.test.js
@@ -3,7 +3,7 @@ import React from 'react'
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
 import { render } from '../../../../test/helpers/render'
-import StreetMetaWidthLabel from '../StreetMetaWidthLabel'
+import StreetMetaWidthLabel from './StreetMetaWidthLabel'
 
 const dummyStreetObject = {
   units: 0,
