+++ assets/scripts/streets/StreetMeta/StreetMetaWidthLabel.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import { FormattedMessage, useIntl } from 'react-intl'
 import { ChevronDownIcon, RulerHorizontalIcon } from '@radix-ui/react-icons'
-import { useSelector } from '../store/hooks'
-import Tooltip from '../ui/Tooltip'
-import { prettifyWidth } from '../util/width_units'
+import { useSelector } from '../../store/hooks'
+import Tooltip from '../../ui/Tooltip'
+import { prettifyWidth } from '../../util/width_units'
 import type { StreetJsonExtra } from '@streetmix/types'
 import './StreetMetaWidthLabel.scss'
