+++ assets/scripts/streets/StreetName.test.js
@@ -2,8 +2,8 @@
 import React from 'react'
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { render } from '../../../../test/helpers/render'
-import StreetName from '../StreetName'
+import { render } from '../../../test/helpers/render'
+import StreetName from './StreetName'
 
 describe('StreetName', () => {
   it('renders a name', () => {
