+++ assets/scripts/streets/thumbnail.test.js
